[ADF-5149] DataTableHeader created and applied transclusion through DataTable
