+++ lib/core/datatable/components/datatable/datatable.component.spec.ts
@@ -160,98 +160,6 @@ describe('DataTable', () => {
         expect(element.querySelector('.adf-datatable-list')).not.toBeNull();
     });
 
-    it('should hide the header if showHeader is false', () => {
-        const newData = new ObjectDataTableAdapter(
-            [
-                { name: '1' },
-                { name: '2' }
-            ],
-            [new ObjectDataColumn({ key: 'name' })]
-        );
-
-        dataTable.showHeader = false;
-        dataTable.loading = false;
-        dataTable.ngOnChanges({
-            data: new SimpleChange(null, newData, false)
-        });
-
-        fixture.detectChanges();
-
-        expect(element.querySelector('.adf-datatable-header')).toBe(null);
-    });
-
-    it('should hide the header if there are no elements inside', () => {
-        const newData = new ObjectDataTableAdapter(
-        );
-
-        dataTable.ngOnChanges({
-            data: new SimpleChange(null, newData, false)
-        });
-
-        fixture.detectChanges();
-
-        expect(element.querySelector('.adf-datatable-header')).toBe(null);
-    });
-
-    it('should hide the header if noPermission is true', () => {
-        const newData = new ObjectDataTableAdapter(
-        );
-
-        dataTable.noPermission = true;
-        dataTable.loading = false;
-
-        dataTable.ngOnChanges({
-            data: new SimpleChange(null, newData, false)
-        });
-
-        fixture.detectChanges();
-
-        expect(element.querySelector('.adf-datatable-header')).toBe(null);
-    });
-
-    it('should show the header if showHeader is true', () => {
-        const newData = new ObjectDataTableAdapter(
-            [
-                { name: '1' },
-                { name: '2' }
-            ],
-            [new ObjectDataColumn({ key: 'name' })]
-        );
-        dataTable.showHeader = true;
-        dataTable.loading = false;
-
-        dataTable.ngOnChanges({
-            data: new SimpleChange(null, newData, false)
-        });
-
-        fixture.detectChanges();
-
-        expect(element.querySelector('.adf-datatable-header')).toBeDefined();
-    });
-
-    it('should emit "sorting-changed" DOM event', (done) => {
-        const column = new ObjectDataColumn({ key: 'name', sortable: true, direction: 'asc' });
-        dataTable.data = new ObjectDataTableAdapter(
-            [
-                { name: '1' },
-                { name: '2' }
-            ],
-            [column]
-        );
-        dataTable.data.setSorting(new DataSorting('name', 'desc'));
-
-        fixture.nativeElement.addEventListener('sorting-changed', (event: CustomEvent) => {
-            expect(event.detail.key).toBe('name');
-            expect(event.detail.direction).toBe('asc');
-            done();
-        });
-
-        dataTable.ngOnChanges({});
-        fixture.detectChanges();
-        dataTable.ngAfterViewInit();
-        dataTable.onColumnHeaderClick(column);
-    });
-
     it('should change the rows on changing of the data', () => {
         const newData = new ObjectDataTableAdapter(
             [
@@ -752,121 +660,6 @@ describe('DataTable', () => {
         expect(e.preventDefault).toHaveBeenCalled();
     });
 
-    it('should not sort if column is missing', () => {
-        dataTable.ngOnChanges({ 'data': new SimpleChange('123', {}, true) });
-        fixture.detectChanges();
-        dataTable.ngAfterViewInit();
-        const adapter = dataTable.data;
-        spyOn(adapter, 'setSorting').and.callThrough();
-        dataTable.onColumnHeaderClick(null);
-        expect(adapter.setSorting).not.toHaveBeenCalled();
-    });
-
-    it('should not sort upon clicking non-sortable column header', () => {
-        dataTable.ngOnChanges({ 'data': new SimpleChange('123', {}, true) });
-        fixture.detectChanges();
-        dataTable.ngAfterViewInit();
-        const adapter = dataTable.data;
-        spyOn(adapter, 'setSorting').and.callThrough();
-
-        const column = new ObjectDataColumn({
-            key: 'column_1'
-        });
-
-        dataTable.onColumnHeaderClick(column);
-        expect(adapter.setSorting).not.toHaveBeenCalled();
-    });
-
-    it('should set sorting upon column header clicked', () => {
-        dataTable.ngOnChanges({ 'data': new SimpleChange('123', {}, true) });
-        fixture.detectChanges();
-        dataTable.ngAfterViewInit();
-        const adapter = dataTable.data;
-        spyOn(adapter, 'setSorting').and.callThrough();
-
-        const column = new ObjectDataColumn({
-            key: 'column_1',
-            sortable: true
-        });
-
-        dataTable.onColumnHeaderClick(column);
-        expect(adapter.setSorting).toHaveBeenCalledWith(
-            jasmine.objectContaining({
-                key: 'column_1',
-                direction: 'asc'
-            })
-        );
-    });
-
-    it('should invert sorting upon column header clicked', () => {
-        dataTable.ngOnChanges({ 'data': new SimpleChange('123', {}, true) });
-        fixture.detectChanges();
-        dataTable.ngAfterViewInit();
-
-        const adapter = dataTable.data;
-        const sorting = new DataSorting('column_1', 'asc');
-        spyOn(adapter, 'setSorting').and.callThrough();
-        spyOn(adapter, 'getSorting').and.returnValue(sorting);
-
-        const column = new ObjectDataColumn({
-            key: 'column_1',
-            sortable: true
-        });
-
-        // check first click on the header
-        dataTable.onColumnHeaderClick(column);
-        expect(adapter.setSorting).toHaveBeenCalledWith(
-            jasmine.objectContaining({
-                key: 'column_1',
-                direction: 'desc'
-            })
-        );
-
-        // check second click on the header
-        sorting.direction = 'desc';
-        dataTable.onColumnHeaderClick(column);
-        expect(adapter.setSorting).toHaveBeenCalledWith(
-            jasmine.objectContaining({
-                key: 'column_1',
-                direction: 'asc'
-            })
-        );
-   });
-
-    it('should indicate column that has sorting applied', () => {
-        dataTable.data = new ObjectDataTableAdapter(
-            [{ name: '1' }, { name: '2' }],
-            [
-                new ObjectDataColumn({ key: 'name', sortable: true }),
-                new ObjectDataColumn({ key: 'other', sortable: true })
-            ]
-        );
-        fixture.detectChanges();
-        dataTable.ngAfterViewInit();
-
-        const [col1, col2] = dataTable.getSortableColumns();
-
-        dataTable.onColumnHeaderClick(col2);
-
-        expect(dataTable.isColumnSortActive(col1)).toBe(false);
-        expect(dataTable.isColumnSortActive(col2)).toBe(true);
-    });
-
-    it('should return false for columns that have no sorting', () => {
-        dataTable.data = new ObjectDataTableAdapter(
-            [{ name: '1' }, { name: '2' }],
-            [
-                new ObjectDataColumn({ key: 'name', sortable: false }),
-                new ObjectDataColumn({ key: 'other', sortable: false })
-            ]
-        );
-
-        const [col1, col2] = dataTable.getSortableColumns();
-
-        expect(dataTable.isColumnSortActive(col1)).toBe(false);
-        expect(dataTable.isColumnSortActive(col2)).toBe(false);
-    });
-
     it('should invert "select all" status', () => {
         expect(dataTable.isSelectAllChecked).toBeFalsy();
         dataTable.onSelectAllClick(<MatCheckboxChange> { checked: true });
@@ -1303,35 +1096,6 @@ describe('Accesibility', () => {
         expect(datatableBodyAttributes.getNamedItem('role').value).toEqual('rowgroup');
         expect(datatableBodyRowAttributes.getNamedItem('role').value).toEqual('row');
         expect(datatableBodyCellAttributes.getNamedItem('role').value).toEqual('gridcell');
-    });
-
-    describe('aria-sort', () => {
-        let column: DataColumn;
-
-        beforeEach(() => {
-            column = new ObjectDataColumn({ key: 'key' });
-        });
-
-        it('should return correct translation key when no sort is applied', () => {
-            spyOn(dataTable, 'isColumnSortActive').and.returnValue(false);
-            expect(dataTable.getAriaSort(column)).toBe('ADF-DATATABLE.ACCESSIBILITY.SORT_NONE');
-        });
-
-        it('should return translation key when column sort is ascending', () => {
-            const isColumnSortedAsc = true;
-            spyOn(dataTable, 'isColumnSortActive').and.returnValue(true);
-            spyOn(dataTable, 'isColumnSorted').and.returnValue(isColumnSortedAsc);
-
-            expect(dataTable.getAriaSort(column)).toBe('ADF-DATATABLE.ACCESSIBILITY.SORT_ASCENDING');
-        });
-
-        it('should return translation key when column sort is descending', () => {
-            const isColumnSortedAsc = false;
-            spyOn(dataTable, 'isColumnSortActive').and.returnValue(true);
-            spyOn(dataTable, 'isColumnSorted').and.returnValue(isColumnSortedAsc);
-
-            expect(dataTable.getAriaSort(column)).toBe('ADF-DATATABLE.ACCESSIBILITY.SORT_DESCENDING');
-        });
     });
 
     it('should focus next row on ArrowDown event', () => {
